backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
Pillow==10.1.0
aiofiles==23.2.1
openai==0.28.1
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
